Fix WS parser: check max_message_size before payload alloc; ssize_t return
Agent-Logs-Url: https://github.com/Coldwings/Elio/sessions/c16c558b-4926-4c4b-81d5-3ee7c5a5efac

+++ include/elio/http/websocket_server.hpp
@@ -198,7 +198,7 @@ public:
                 co_return std::nullopt;
             }
 
-            int parsed = parser_.parse(
+            ssize_t parsed = parser_.parse(
                 reinterpret_cast<const uint8_t*>(buffer_.data()),
                 static_cast<size_t>(result.result)
             );
